Make helper functions for account and profile creation.
